+++ kernel/src/irq.c
@@ -1,6 +1,7 @@
 #include <x86/pic.h>
 #include <x86/idt.h>
 #include <kernel/irq.h>
+#include <kernel/bsp.h>
 
 #define IRQ_OFFSET 0x20
 
@@ -11,5 +12,5 @@ irq_init(void) {
 
 void
 irq_set_handler(uint8_t number, uint8_t ist, int_handler handler) {
-  idt_set_int_handler(IRQ_OFFSET+number, ist, handler);
+  idt_set_int_handler((uint8_t)(IRQ_OFFSET+number), ist, handler);
 }
